+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - Integrated Connection template actions in Ethernet network within the update action, and added `:reset_connection_template` action
   - Added support to Logical enclosure `:reconfigure` and `:set_script`
   - Added support to Network set actions
+  - Added support to Datacenter actions
 
 ### 0.1.1
   - Fixed Ruby SDK version to 1.0.0